refactor: apiUrl, queryKeys, useCategory hook 폴더 위치 변경
