+++ src/hooks/queries/useCategory.ts
@@ -1,5 +1,5 @@
 import { getCategory } from '@/api/category';
-import { CategoryQueryKeys } from '@/utils/queryKeys';
+import { CategoryQueryKeys } from '@/constants/queryKeys';
 import { useQuery } from '@tanstack/react-query';
 
 export function useCategory() {
